+++ drivers/cc110x/cc110x-rxtx.c
@@ -24,10 +24,10 @@
 #include <string.h>
 
 #include "cc110x.h"
-#include "cc110x/cc110x-spi.h"
-#include "cc110x/cc110x-internal.h"
-#include "cc110x/cc110x-interface.h"
-#include "cc110x/cc110x-defines.h"
+#include "cc110x-spi.h"
+#include "cc110x-internal.h"
+#include "cc110x-interface.h"
+#include "cc110x-defines.h"
 
 #include "periph/gpio.h"
 #include "irq.h"
