+++ drivers/cc110x/cc110x.c
@@ -29,11 +29,11 @@
 #include "log.h"
 
 #include "cc110x.h"
-#include "cc110x/cc110x-defaultsettings.h"
-#include "cc110x/cc110x-defines.h"
-#include "cc110x/cc110x-interface.h"
-#include "cc110x/cc110x-internal.h"
-#include "cc110x/cc110x-spi.h"
+#include "cc110x-defaultsettings.h"
+#include "cc110x-defines.h"
+#include "cc110x-interface.h"
+#include "cc110x-internal.h"
+#include "cc110x-spi.h"
 
 #define ENABLE_DEBUG    (0)
 #include "debug.h"
