+++ drivers/cc110x/gnrc_cc110x/gnrc_netdev2_cc110x.c
@@ -13,7 +13,7 @@
 #include "net/netdev2.h"
 #include "net/gnrc.h"
 #include "cc110x.h"
-#include "cc110x/cc110x-netdev2.h"
+#include "cc110x-netdev2.h"
 #include "net/gnrc/gnrc_netdev2.h"
 #include "od.h"
