+++ drivers/cc110x/include/cc110x-defaultsettings.h
