+++ drivers/cc110x/include/cc110x-defines.h
