+++ drivers/cc110x/include/cc110x-spi.h
